Remove unused knowledge bases
These will not be used by the final system.

You can still take a look at the examples, e.g. in the original
repository jelmervdl/kennissysteem or in the diff for this commit.  In
particular, knowledge.xml is already a very simple example for a
building safety-knowledge base.  However, it is written in Dutch.