Remove debug-log and all references to it
The Solver-class has a log that is only displayed in the debug-panel
which was already removed. Remove the log, the corresponding log-method
and all its calls as well as the now unused WebLogger class,
Logger-interface and the use of the WebLogger in webfrontend.php.

This fixes an error that occured when using the CLI-version, too.