Remove v-option of the CLI-version
The verbose-option of the CLI-version does not work anyway, so all
references to it can be removed.

The original repository made the v-option functional by replacing the
verbose-function by a Logger and making the Logger-argument of the
Solver-constructor optional. However, all logging in this repository was
removed. If a verbose-option is needed for debugging, it can be
reimplemented using a Logger. In any case, the verbose-function will not
be needed.