Improve error-handling in process_file-function
This index.php-function had return values of different types with some
errors being stored in errors[] and some being returned. The returned
errors were then erroneously used as file names.
Now the function has no return-value anymore. All occurring errors are
stored in errors[] with four keys:
    'number'  => error level
    'message' => description of error
    'file'    => file in which error occurred
    'line'    => line in which error occurred

This display of the errors in single.phtml was changed such that the
file name and line number are actually shown. Before, the
title-attribute was used although the li-tag does not have such an
attribute.