Restructure knowledgebase-xml-documents
Replaces goals with answers by buildings and checklist_items in
knowledge.dtd.

Buildings have a name-attribute, a value-attribute and multiple risks
that are associated with the building. The name of the building
corresponds to a fact-name and its value to the content of the fact-tag.
If the fact with name is deduced and has the correct content
(=building-value), we assume that the building is one that the user is
talking about. If the fact with name is deduced and has incorrect
content, we assume that the building is not one that the user is talking
about. The risks of the deduced buildings then tell us which
checklist_items will be used.

Checklist_items have a name that corresponds to the name of a risk
associated with a building (e.g. "hand-burn"). They also have a
description, typically a question that asks whether the risk occurs
("Could a person burn his hands in your building?") and advice that is
given ("Restrict access to the stove plates.") to prevent the possible
danger.