Restructure knowledgebase-reading in Webfrontend
Since the parsing of a knowledgebase-file now results in a
KnowledgeState and a Checklist, both are stored seperately.

Each of them is either decoded from a hidden input tag or read from the
file. Buildings are only pushed to the KnowledgeState's goalStack if the
KnowledgeState was read from the file.