Implement the Checklist-processing in Webfrontend
After all questions are asked, the used template is checklist.phtml.  If
the radioGroup-array is then set (i.e. the checklist was filled out),
the risk_factors and potential_injuries for the ChecklistItems are set.
The security_levels are calculated, the Checklist is sorted, and the
completed.phtml-template is rendered.