Add checklist.phtml-template
The template is integrated in the Webfrontend such that the user input
is processed correctly (the ChecklistItems will have a set risk_factor
and potential_injury and are sorted depending on their security_levels).

It showcases the basic functionality of the implemented checklist and
will be improved upon in the future.