Implement getDeducedBuildings() in KnowledgeState
It is not the responsibility of the Checklist to check which buildings
were deduced from the facts. It is the responsibility of the
KnowledgeState itself. To this end, use the new function
getDeducedBuildings().

The function is used in webfrontend.php 1. to give the deduced buildings
to the checklist.phtml-Template such that they can be displayed there,
and 2. to give them to the Checklist's create-function which now does
not take a KnowledgeState anymore but an IteratorAggregate of Buildings.
Using these Buildings, the Checklist restructures itself such that it
only contains the ChecklistItems that belong to the given Buildings.