Sort ChecklistItems by description
In checklist.phtml the ChecklistItems should be sorted by location (part
of a Building, e.g. playroom, kitchen etc.).  However, a location
attribute for ChecklistItems is not implemented so far.  Instead, the
location is mentioned in the ChecklistItem-description using the format
"Location: Question?" and the ChecklistItems are sorted by description.