Remove AskedQuestion-class
The class should have been called QuestionToBeAsked as it is returned
after the Solver determined which Question is to be asked next.

The only functionality it adds is that Questions can be marked as
skippable when other Questions/Rules can be used to infer the same
thing. Then an "I don't know"-option is added to the question. This can
lead to weird situations, however, where questions are first skipped and
then occur again. In our system questions should not be skipped anyway.