Move analyse-facts-overview to a single table
Remove some redundant/unnecessary parts of the css as well.

It looks much cleaner when all of the fact-data is in a single table.
Moreover, it wastes less space.
However, it might be nice to still see the headers after scrolling down.
A potential solution would be a table with a fixed header and scrollable
body, but this is very difficult to implement. In the future, the
headers might just be repeated for every fact.